Add proper support for most digitalCAT flags
According to the digitalCAT user guide, there are 8 flags:
- Prefix
- Suffix
- Capitalize Next
- Attach Letters
- Attach Numbers
- Double Last
- OR Ending
- Proper Name

Inspecting various dictionaries yields the actual flag values for each
flag, which appear to be in the same order.  "OR Ending", "Double Last",
and "Proper Name" are unnecessary in Plover, but this change implements
the missing flags.

An example of some "Capitalize Next" entries given in the user guide
include entries such as `Mr.^` and `St.^`, while inspecting dictionaries
suggests entries such as `Ms. ^` are also included.  As such a strip is
added to account for the difference in space handling between digitalCAT
and Plover.